avoid floating point drift

+++ 04-1DReduction/reduce.metal
@@ -4,9 +4,9 @@ using namespace metal;
 // cf. https://github.com/NVIDIA/cuda-samples/blob/master/Samples/2_Concepts_and_Techniques/reduction/reduction_kernel.cu
 
 
-kernel void reduceSum1D_2(device const float* X [[buffer(0)]],
+kernel void reduceSum1D_2(device const int* X [[buffer(0)]],
                           device const unsigned long *xLength [[buffer(1)]],
-                          device float* result [[buffer(2)]],
+                          device long* result [[buffer(2)]],
                           threadgroup float* shared_data [[threadgroup(0)]],
                           uint global_id [[thread_position_in_grid]],
                           uint group_id [[threadgroup_position_in_grid]],
@@ -17,7 +17,7 @@ kernel void reduceSum1D_2(device const float* X [[buffer(0)]],
     uint i = global_id;
     uint n = xLength[0]; // Note: the [[grid_size]] attribute is not set for dispatchThreadgroups()
 
-    shared_data[tid] = (i < n) ? X[i] : 0.0;
+    shared_data[tid] = (i < n) ? X[i] : 0;
     threadgroup_barrier(mem_flags::mem_threadgroup);
 
     // do reduction in shared memory
